Treat a failed dill import the same as dill not existing at all
Prior to importing `dill`, the import system is used to ask if it can
find if its potentially importable. This probing essentially just
verifies if the top-level `dill/__init__.py` file exists, but doesn't
verify that the entire dill package is actually importable and
available. While this is reliable in a typical environment, it doesn't
work well when there are user-installed custom importers, creative file
system layours, or esoteric sys.path modifications.

To fix, simply handle a failing dill import the same as if dill didn't
appear to be available at all.